refactor: element imports field/pool and field/hash to avoid cycles
