+++ field/hash/hashutils_test.go
@@ -1,4 +1,4 @@
-package field
+package hash
 
 import (
 	"bytes"
